+++ CHANGELOG.md
@@ -38,6 +38,8 @@ certain VM instructions are used.
 shallow comparison was performed, so it was working just for plain values such as atoms and small
 integers
 - Fixed support for setting esp32 boot_path in NVS.
+- Fixed race conditions in network:start/stop.
+- Fixed crash calling network:sta_rssi(), when network not up.
 
 ## [0.6.5] - 2024-10-15